test(sessions): split redaction cases out of session-manager.test.ts
session-manager.test.ts crossed the repo's 1000-line oxlint cap after the
transcript-append refusal redaction tests landed. Move the three new
redaction/hash cases (collision-hash boundary, newline/ANSI/oversized
identifier, and the operator-facing compaction reason) to a sibling file,
following the same split pattern already used for
session-manager.fork-rebase.test.ts and agent-exec.default-agent.test.ts
(#119765). No assertions were weakened, only relocated.
